Modularize attachment backup API routes
Separate ledger transport and wire models from HTTP handlers, and group endpoint scenarios into focused test modules.
