+++ crates/api-sync/src/routes/attachment_backups.rs
@@ -1,5 +1,3 @@
-use std::time::Duration;
-
 use anlg_api_auth::AuthContext;
 use axum::{
     Extension, Json, Router,
@@ -10,10 +8,7 @@ use axum::{
     routing::{get, post, put},
 };
 use chrono::Utc;
-#[cfg(test)]
-use chrono::{SecondsFormat, TimeDelta};
-use reqwest::StatusCode;
-use serde::{Deserialize, Serialize, de::DeserializeOwned};
+use serde::{Deserialize, Serialize};
 use utoipa::OpenApi;
 
 use crate::{
@@ -21,20 +16,28 @@ use crate::{
     state::AppState,
 };
 
+mod ledger;
 mod validation;
 
+#[cfg(test)]
+use ledger::MAX_BACKUP_RPC_RESPONSE_BYTES;
+use ledger::{
+    CanceledDeletionRow, CurrentRow, CurrentRpcRequest, DeleteRpcRequest, FinalizeRpcRequest,
+    FinalizedRow, MarkSignedRpcRequest, MarkedSignedRow, PrepareDownloadRpcRequest,
+    PreparedDownloadRow, PromoteRpcRequest, PromotedRow, ReserveRpcRequest, ReservedRow,
+    ScheduledDeletionRow, map_backup_error, map_cancel_deletion_error, map_deletion_error,
+    map_reservation_error, read_backup_object, rpc_single,
+};
 use validation::{
     canonical_object_id, canonical_owner, future_timestamp, invalid_request,
-    invalid_upstream_response, parse_timestamp, require_pro, valid_size,
-    validate_backup_object_row, validate_delete_request, validate_deletion_row_identity,
-    validate_marked_signed_row, validate_object_key, validate_prepared_download,
-    validate_promoted_row, validate_ref, validate_reserved_row, validate_sha256,
+    invalid_upstream_response, parse_timestamp, require_pro, valid_size, validate_delete_request,
+    validate_deletion_row_identity, validate_marked_signed_row, validate_object_key,
+    validate_prepared_download, validate_promoted_row, validate_ref, validate_reserved_row,
+    validate_sha256,
 };
 
 const ATTACHMENT_BACKUP_BUCKET: &str = "attachment-backups";
-const BACKUP_RPC_TIMEOUT: Duration = Duration::from_secs(5);
 const MAX_BACKUP_REQUEST_BYTES: usize = 4 * 1024;
-const MAX_BACKUP_RPC_RESPONSE_BYTES: usize = 16 * 1024;
 const MAX_CIPHERTEXT_SIZE_BYTES: u64 = 545_259_520;
 const FORMAT_VERSION: i16 = 1;
 const UPLOAD_TOKEN_TTL_SECONDS: i64 = 2 * 60 * 60;
@@ -165,179 +168,6 @@ pub(super) struct CanceledAttachmentBackupDeletion {
     attachment_ref: String,
     version_ref: String,
     delete_request_id: String,
-}
-
-#[derive(Serialize)]
-struct ReserveRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_attachment_ref: &'a str,
-    p_version_ref: &'a str,
-    p_ciphertext_size_bytes: i64,
-    p_format_version: i16,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ReservedRow {
-    object_id: String,
-    object_key: String,
-    object_state: String,
-    ciphertext_size_bytes: i64,
-    format_version: i16,
-    reservation_expires_at: String,
-    cleanup_not_before: String,
-    ciphertext_sha256: Option<String>,
-    was_created: bool,
-}
-
-#[derive(Serialize)]
-struct MarkSignedRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_object_id: &'a str,
-    p_upload_expires_at: &'a str,
-    p_ciphertext_sha256: &'a str,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct MarkedSignedRow {
-    object_id: String,
-    object_key: String,
-    last_signed_at: String,
-    upload_expires_at: String,
-    cleanup_not_before: String,
-    ciphertext_sha256: String,
-}
-
-#[derive(Serialize)]
-struct ObjectKeyRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_object_key: &'a str,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct BackupObjectRow {
-    object_id: String,
-    attachment_ref: String,
-    version_ref: String,
-    object_key: String,
-    object_state: String,
-    ciphertext_size_bytes: i64,
-    format_version: i16,
-    reservation_expires_at: String,
-    upload_expires_at: Option<String>,
-    cleanup_not_before: String,
-    ciphertext_sha256: Option<String>,
-}
-
-#[derive(Serialize)]
-struct FinalizeRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_object_id: &'a str,
-    p_object_key: &'a str,
-    p_observed_ciphertext_size_bytes: i64,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct FinalizedRow {
-    object_id: String,
-    object_key: String,
-    object_state: String,
-    was_finalized: bool,
-}
-
-#[derive(Serialize)]
-struct PromoteRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_candidate_object_id: &'a str,
-    p_candidate_object_key: &'a str,
-    p_expected_current_object_key: Option<&'a str>,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct PromotedRow {
-    current_object_id: String,
-    current_object_key: String,
-    current_version_ref: String,
-    current_ciphertext_sha256: String,
-    displaced_object_id: Option<String>,
-    displaced_object_key: Option<String>,
-    was_promoted: bool,
-}
-
-#[derive(Serialize)]
-struct CurrentRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_attachment_ref: &'a str,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct CurrentRow {
-    object_id: String,
-    version_ref: String,
-    object_key: String,
-    ciphertext_sha256: String,
-    ciphertext_size_bytes: i64,
-    format_version: i16,
-}
-
-#[derive(Serialize)]
-struct PrepareDownloadRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_object_key: &'a str,
-    p_download_expires_at: &'a str,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct PreparedDownloadRow {
-    object_id: String,
-    object_key: String,
-    ciphertext_sha256: String,
-    ciphertext_size_bytes: i64,
-    format_version: i16,
-    cleanup_not_before: String,
-}
-
-#[derive(Serialize)]
-struct DeleteRpcRequest<'a> {
-    p_owner_user_id: &'a str,
-    p_attachment_ref: &'a str,
-    p_version_ref: &'a str,
-    p_object_key: &'a str,
-    p_delete_request_id: &'a str,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ScheduledDeletionRow {
-    outcome: String,
-    object_id: Option<String>,
-    object_key: String,
-    delete_request_id: String,
-    delete_fence_id: Option<String>,
-    delete_generation: Option<i64>,
-    delete_not_before: Option<String>,
-    #[serde(rename = "was_created")]
-    _was_created: bool,
-}
-
-#[derive(Deserialize)]
-#[serde(deny_unknown_fields)]
-struct CanceledDeletionRow {
-    outcome: String,
-    object_key: String,
-    #[serde(rename = "was_cancelled")]
-    _was_cancelled: bool,
-}
-
-#[derive(Deserialize)]
-struct PostgrestError {
-    code: String,
 }
 
 #[derive(OpenApi)]
@@ -977,1645 +807,5 @@ async fn cancel_attachment_backup_deletion(
     }))
 }
 
-#[derive(Debug)]
-enum RpcFailure {
-    Empty,
-    Rejected {
-        status: StatusCode,
-        code: Option<String>,
-    },
-    Unavailable,
-}
-
-async fn rpc_single<RequestBody, Row>(
-    state: &AppState,
-    function: &str,
-    request: &RequestBody,
-) -> std::result::Result<Row, RpcFailure>
-where
-    RequestBody: Serialize + ?Sized,
-    Row: DeserializeOwned,
-{
-    let mut response = state
-        .client
-        .post(format!(
-            "{}/rest/v1/rpc/{function}",
-            state.config.supabase_url
-        ))
-        .header("apikey", &state.config.supabase_service_role_key)
-        .bearer_auth(&state.config.supabase_service_role_key)
-        .timeout(BACKUP_RPC_TIMEOUT)
-        .json(request)
-        .send()
-        .await
-        .map_err(|_| {
-            tracing::warn!(function, "Attachment backup ledger request failed");
-            RpcFailure::Unavailable
-        })?;
-    let status = response.status();
-    if response
-        .content_length()
-        .is_some_and(|length| length > MAX_BACKUP_RPC_RESPONSE_BYTES as u64)
-    {
-        tracing::warn!(function, %status, "Attachment backup ledger response was too large");
-        return Err(RpcFailure::Unavailable);
-    }
-
-    let mut bytes = Vec::new();
-    while let Some(chunk) = response.chunk().await.map_err(|_| {
-        tracing::warn!(
-            function,
-            "Attachment backup ledger response could not be read"
-        );
-        RpcFailure::Unavailable
-    })? {
-        if bytes.len().saturating_add(chunk.len()) > MAX_BACKUP_RPC_RESPONSE_BYTES {
-            tracing::warn!(function, %status, "Attachment backup ledger response was too large");
-            return Err(RpcFailure::Unavailable);
-        }
-        bytes.extend_from_slice(&chunk);
-    }
-
-    if !status.is_success() {
-        let code = serde_json::from_slice::<PostgrestError>(&bytes)
-            .ok()
-            .map(|error| error.code);
-        tracing::warn!(function, %status, ?code, "Attachment backup ledger request was rejected");
-        return Err(RpcFailure::Rejected { status, code });
-    }
-
-    let mut rows = serde_json::from_slice::<Vec<Row>>(&bytes).map_err(|_| {
-        tracing::warn!(function, "Attachment backup ledger response was invalid");
-        RpcFailure::Unavailable
-    })?;
-    match rows.len() {
-        0 => Err(RpcFailure::Empty),
-        1 => Ok(rows.pop().expect("row count was checked")),
-        row_count => {
-            tracing::warn!(
-                function,
-                row_count,
-                "Attachment backup ledger returned multiple rows"
-            );
-            Err(RpcFailure::Unavailable)
-        }
-    }
-}
-
-async fn read_backup_object(
-    state: &AppState,
-    owner_user_id: &str,
-    object_key: &str,
-) -> Result<BackupObjectRow> {
-    let row: BackupObjectRow = rpc_single(
-        state,
-        "read_attachment_backup_by_key",
-        &ObjectKeyRpcRequest {
-            p_owner_user_id: owner_user_id,
-            p_object_key: object_key,
-        },
-    )
-    .await
-    .map_err(map_backup_error)?;
-    validate_backup_object_row(&row, owner_user_id, object_key)?;
-    Ok(row)
-}
-
-fn map_reservation_error(error: RpcFailure) -> SyncError {
-    match &error {
-        RpcFailure::Rejected { code, .. } if code.as_deref() == Some("54000") => {
-            SyncError::AttachmentBackupQuotaExceeded
-        }
-        _ => map_backup_error(error),
-    }
-}
-
-fn map_deletion_error(error: RpcFailure) -> SyncError {
-    match &error {
-        RpcFailure::Rejected { code, .. } if code.as_deref() == Some("55000") => {
-            SyncError::AttachmentBackupNotFound
-        }
-        _ => map_backup_error(error),
-    }
-}
-
-fn map_cancel_deletion_error(error: RpcFailure) -> SyncError {
-    match &error {
-        RpcFailure::Rejected { code, .. } if code.as_deref() == Some("55006") => {
-            SyncError::AttachmentBackupDeleteTooLate
-        }
-        _ => map_backup_error(error),
-    }
-}
-
-fn map_backup_error(error: RpcFailure) -> SyncError {
-    match error {
-        RpcFailure::Empty => SyncError::AttachmentBackupNotFound,
-        RpcFailure::Rejected { status, code }
-            if status == StatusCode::UNAUTHORIZED
-                || status == StatusCode::FORBIDDEN
-                || code.as_deref() == Some("42501") =>
-        {
-            SyncError::AttachmentBackupForbidden
-        }
-        RpcFailure::Rejected { code, .. } if code.as_deref() == Some("22023") => invalid_request(),
-        RpcFailure::Rejected { code, .. } if matches!(code.as_deref(), Some("40001" | "55000")) => {
-            SyncError::AttachmentBackupConflict
-        }
-        RpcFailure::Rejected { .. } | RpcFailure::Unavailable => {
-            SyncError::AttachmentBackupServiceUnavailable
-        }
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use anlg_api_auth::{AuthContext, Claims};
-    use axum::{
-        body::{Body, to_bytes},
-        http::{Method, Request, StatusCode, header},
-    };
-    use serde_json::{Value, json};
-    use tower::ServiceExt;
-    use wiremock::{
-        Mock, MockServer, ResponseTemplate,
-        matchers::{body_partial_json, header as request_header, method, path},
-    };
-
-    use super::*;
-    use crate::SyncConfig;
-
-    const OWNER: &str = "11111111-1111-4111-8111-111111111111";
-    const OBJECT_ID: &str = "22222222-2222-4222-8222-222222222222";
-    const OBJECT_UUID: &str = "33333333-3333-4333-8333-333333333333";
-    const DELETE_REQUEST_ID: &str = "44444444-4444-4444-8444-444444444444";
-    const DISPLACED_UUID: &str = "55555555-5555-4555-8555-555555555555";
-    const DELETE_FENCE_ID: &str = "66666666-6666-4666-8666-666666666666";
-    const ATTACHMENT_REF: &str = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA";
-    const VERSION_REF: &str = "BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB";
-    const CIPHERTEXT_SHA256: &str =
-        "ad47fd9e87159d651a53b3dfba3ef200684a9ed88c2528b62e18f3881fe203b0";
-
-    fn object_key() -> String {
-        format!("{OWNER}/{OBJECT_UUID}.anb1")
-    }
-
-    fn displaced_key() -> String {
-        format!("{OWNER}/{DISPLACED_UUID}.anb1")
-    }
-
-    fn timestamp_after(seconds: i64) -> String {
-        (Utc::now() + TimeDelta::seconds(seconds)).to_rfc3339_opts(SecondsFormat::Secs, true)
-    }
-
-    fn test_state(server: &MockServer) -> AppState {
-        AppState::new(SyncConfig {
-            project_url: server.uri(),
-            token_issuer_api_key: "issuer-key".to_string(),
-            database_id: "database-id".to_string(),
-            legacy_database_id: None,
-            protocol_mode: crate::config::CloudsyncProtocolMode::E2eeEnforced,
-            token_ttl_seconds: 60,
-            supabase_url: server.uri(),
-            supabase_anon_key: "anon-key".to_string(),
-            supabase_service_role_key: "service-role-key".to_string(),
-        })
-    }
-
-    fn test_router_with_state(state: AppState, is_pro: bool) -> axum::Router {
-        router().with_state(state).layer(Extension(AuthContext {
-            token: "user-token".to_string(),
-            claims: Claims {
-                sub: OWNER.to_string(),
-                email: None,
-                entitlements: is_pro
-                    .then(|| "hyprnote_pro".to_string())
-                    .into_iter()
-                    .collect(),
-                subscription_status: None,
-                trial_end: None,
-                has_payment_method: None,
-            },
-        }))
-    }
-
-    fn test_router(server: &MockServer, is_pro: bool) -> axum::Router {
-        test_router_with_state(test_state(server), is_pro)
-    }
-
-    fn json_request(method: Method, path: &str, body: Value) -> Request<Body> {
-        Request::builder()
-            .method(method)
-            .uri(path)
-            .header(header::CONTENT_TYPE, "application/json")
-            .body(Body::from(serde_json::to_vec(&body).unwrap()))
-            .unwrap()
-    }
-
-    async fn response_json(response: axum::response::Response) -> Value {
-        let bytes = to_bytes(response.into_body(), 64 * 1024).await.unwrap();
-        serde_json::from_slice(&bytes).unwrap()
-    }
-
-    fn reserved_row(state: &str, hash: Option<&str>) -> Value {
-        json!({
-            "object_id": OBJECT_ID,
-            "object_key": object_key(),
-            "object_state": state,
-            "ciphertext_sha256": hash,
-            "ciphertext_size_bytes": 1234,
-            "format_version": 1,
-            "reservation_expires_at": timestamp_after(15 * 60),
-            "cleanup_not_before": timestamp_after(30 * 60),
-            "was_created": true
-        })
-    }
-
-    fn object_row(state: &str, hash: Option<&str>) -> Value {
-        json!({
-            "object_id": OBJECT_ID,
-            "attachment_ref": ATTACHMENT_REF,
-            "version_ref": VERSION_REF,
-            "object_key": object_key(),
-            "object_state": state,
-            "ciphertext_sha256": hash,
-            "ciphertext_size_bytes": 1234,
-            "format_version": 1,
-            "reservation_expires_at": timestamp_after(15 * 60),
-            "upload_expires_at": hash.map(|_| timestamp_after(2 * 60 * 60)),
-            "cleanup_not_before": timestamp_after(26 * 60 * 60)
-        })
-    }
-
-    fn delete_request_body() -> Value {
-        json!({
-            "objectKey": object_key(),
-            "attachmentRef": ATTACHMENT_REF,
-            "versionRef": VERSION_REF,
-            "deleteRequestId": DELETE_REQUEST_ID
-        })
-    }
-
-    fn scheduled_deletion_row(delete_not_before: &str, was_created: bool) -> Value {
-        json!({
-            "outcome": "scheduled",
-            "object_id": OBJECT_ID,
-            "object_key": object_key(),
-            "delete_request_id": DELETE_REQUEST_ID,
-            "delete_fence_id": DELETE_FENCE_ID,
-            "delete_generation": 7,
-            "delete_not_before": delete_not_before,
-            "was_created": was_created
-        })
-    }
-
-    async fn mount_rpc(server: &MockServer, function: &str, response: ResponseTemplate) {
-        Mock::given(method("POST"))
-            .and(path(format!("/rest/v1/rpc/{function}")))
-            .and(request_header("apikey", "service-role-key"))
-            .and(request_header("authorization", "Bearer service-role-key"))
-            .respond_with(response)
-            .mount(server)
-            .await;
-    }
-
-    #[tokio::test]
-    async fn reserves_identity_without_issuing_a_storage_capability() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "reserve_attachment_backup",
-            ResponseTemplate::new(200).set_body_json(json!([reserved_row("reserved", None)])),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/reserve",
-                json!({
-                    "attachmentRef": ATTACHMENT_REF,
-                    "versionRef": VERSION_REF,
-                    "ciphertextSizeBytes": 1234,
-                    "formatVersion": 1
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(response.headers()[header::CACHE_CONTROL], "no-store");
-        let body = response_json(response).await;
-        assert_eq!(body["objectId"], OBJECT_ID);
-        assert_eq!(body["ciphertextSha256"], Value::Null);
-        assert!(body.get("uploadToken").is_none());
-        let requests = server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 1);
-        assert_eq!(
-            requests[0].url.path(),
-            "/rest/v1/rpc/reserve_attachment_backup"
-        );
-    }
-
-    #[tokio::test]
-    async fn reports_reservation_races_as_conflicts() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "reserve_attachment_backup",
-            ResponseTemplate::new(409).set_body_json(json!({
-                "code": "40001",
-                "message": "reservation-secret-must-not-leak"
-            })),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/reserve",
-                json!({
-                    "attachmentRef": ATTACHMENT_REF,
-                    "versionRef": VERSION_REF,
-                    "ciphertextSizeBytes": 1234,
-                    "formatVersion": 1
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        let body = response_json(response).await;
-        assert_eq!(body["error"]["code"], "attachment_backup_conflict");
-        assert!(!body.to_string().contains("reservation-secret"));
-    }
-
-    #[tokio::test]
-    async fn reports_reservation_limit_as_conflict() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "reserve_attachment_backup",
-            ResponseTemplate::new(409).set_body_json(json!({
-                "code": "55000",
-                "message": "reservation-limit-secret-must-not-leak"
-            })),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/reserve",
-                json!({
-                    "attachmentRef": ATTACHMENT_REF,
-                    "versionRef": VERSION_REF,
-                    "ciphertextSizeBytes": 1234,
-                    "formatVersion": 1
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        let body = response_json(response).await;
-        assert_eq!(body["error"]["code"], "attachment_backup_conflict");
-        assert!(!body.to_string().contains("reservation-limit-secret"));
-    }
-
-    #[tokio::test]
-    async fn reads_then_marks_integrity_before_signing_upload() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200).set_body_json(json!([object_row("reserved", None)])),
-        )
-        .await;
-        Mock::given(method("POST"))
-            .and(path("/rest/v1/rpc/mark_attachment_backup_signed"))
-            .and(body_partial_json(json!({
-                "p_owner_user_id": OWNER,
-                "p_object_id": OBJECT_ID,
-                "p_ciphertext_sha256": CIPHERTEXT_SHA256
-            })))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!([{
-                "object_id": OBJECT_ID,
-                "object_key": object_key(),
-                "ciphertext_sha256": CIPHERTEXT_SHA256,
-                "last_signed_at": timestamp_after(-1),
-                "upload_expires_at": timestamp_after(2 * 60 * 60 + 30),
-                "cleanup_not_before": timestamp_after(27 * 60 * 60)
-            }])))
-            .mount(&server)
-            .await;
-        Mock::given(method("POST"))
-            .and(path(format!(
-                "/storage/v1/object/upload/sign/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "url": format!(
-                    "/object/upload/sign/{ATTACHMENT_BACKUP_BUCKET}/{}?token=upload-secret",
-                    object_key()
-                )
-            })))
-            .mount(&server)
-            .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/upload-grant",
-                json!({
-                    "objectKey": object_key(),
-                    "ciphertextSha256": CIPHERTEXT_SHA256
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = response_json(response).await;
-        assert_eq!(body["uploadToken"], "upload-secret");
-        assert_eq!(body["ciphertextSha256"], CIPHERTEXT_SHA256);
-
-        let requests = server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 3);
-        assert_eq!(
-            requests[0].url.path(),
-            "/rest/v1/rpc/read_attachment_backup_by_key"
-        );
-        assert_eq!(
-            requests[1].url.path(),
-            "/rest/v1/rpc/mark_attachment_backup_signed"
-        );
-        assert!(
-            requests[2]
-                .url
-                .path()
-                .contains("/storage/v1/object/upload/sign/")
-        );
-    }
-
-    #[tokio::test]
-    async fn never_issues_upload_token_for_an_unsafe_cleanup_window() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200).set_body_json(json!([object_row("reserved", None)])),
-        )
-        .await;
-        mount_rpc(
-            &server,
-            "mark_attachment_backup_signed",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "object_id": OBJECT_ID,
-                "object_key": object_key(),
-                "ciphertext_sha256": CIPHERTEXT_SHA256,
-                "last_signed_at": timestamp_after(-1),
-                "upload_expires_at": timestamp_after(2 * 60 * 60 + 4 * 60),
-                "cleanup_not_before": timestamp_after(26 * 60 * 60 + 6 * 60)
-            }])),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/upload-grant",
-                json!({
-                    "objectKey": object_key(),
-                    "ciphertextSha256": CIPHERTEXT_SHA256
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
-        let requests = server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 2);
-        assert!(
-            !requests
-                .iter()
-                .any(|request| request.url.path().contains("/storage/v1/"))
-        );
-    }
-
-    #[tokio::test]
-    async fn never_issues_upload_token_when_integrity_mark_fails() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200).set_body_json(json!([object_row("reserved", None)])),
-        )
-        .await;
-        mount_rpc(
-            &server,
-            "mark_attachment_backup_signed",
-            ResponseTemplate::new(500).set_body_json(json!({
-                "code": "XX000",
-                "message": "database-secret-must-not-leak"
-            })),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/upload-grant",
-                json!({
-                    "objectKey": object_key(),
-                    "ciphertextSha256": CIPHERTEXT_SHA256
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
-        assert_eq!(response.headers()[header::CACHE_CONTROL], "no-store");
-        let body = response_json(response).await.to_string();
-        assert!(!body.contains("database-secret"));
-        assert!(!body.contains("upload-secret"));
-        let requests = server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 2);
-        assert!(
-            !requests
-                .iter()
-                .any(|request| request.url.path().contains("/storage/v1/"))
-        );
-    }
-
-    #[tokio::test]
-    async fn reports_integrity_mark_races_as_conflicts_without_signing() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200).set_body_json(json!([object_row("reserved", None)])),
-        )
-        .await;
-        mount_rpc(
-            &server,
-            "mark_attachment_backup_signed",
-            ResponseTemplate::new(409).set_body_json(json!({
-                "code": "40001",
-                "message": "hash-secret-must-not-leak"
-            })),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/upload-grant",
-                json!({
-                    "objectKey": object_key(),
-                    "ciphertextSha256": CIPHERTEXT_SHA256
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        let body = response_json(response).await;
-        assert_eq!(body["error"]["code"], "attachment_backup_conflict");
-        assert!(!body.to_string().contains("hash-secret"));
-        assert!(
-            !server
-                .received_requests()
-                .await
-                .unwrap()
-                .iter()
-                .any(|request| request.url.path().contains("/storage/v1/"))
-        );
-    }
-
-    #[tokio::test]
-    async fn verifies_storage_metadata_before_finalizing() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("reserved", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/info/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "size": 1234,
-                "content_type": "application/octet-stream",
-                "metadata": {
-                    "ciphertextSha256": CIPHERTEXT_SHA256,
-                    "formatVersion": 1
-                }
-            })))
-            .mount(&server)
-            .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/authenticated/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(vec![0_u8; 1_234]))
-            .mount(&server)
-            .await;
-        mount_rpc(
-            &server,
-            "finalize_attachment_backup",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "object_id": OBJECT_ID,
-                "object_key": object_key(),
-                "object_state": "ready",
-                "was_finalized": true
-            }])),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(response_json(response).await["objectState"], "ready");
-        let requests = server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 4);
-        assert_eq!(
-            requests[0].url.path(),
-            "/rest/v1/rpc/read_attachment_backup_by_key"
-        );
-        assert!(requests[1].url.path().contains("/storage/v1/object/info/"));
-        assert!(
-            requests[2]
-                .url
-                .path()
-                .contains("/storage/v1/object/authenticated/")
-        );
-        assert_eq!(
-            requests[3].url.path(),
-            "/rest/v1/rpc/finalize_attachment_backup"
-        );
-    }
-
-    #[tokio::test]
-    async fn rejects_expired_finalize_before_storage_verification() {
-        let server = MockServer::start().await;
-        let mut row = object_row("reserved", Some(CIPHERTEXT_SHA256));
-        row["reservation_expires_at"] = json!(timestamp_after(-3 * 60 * 60));
-        row["upload_expires_at"] = json!(timestamp_after(-2 * 60 * 60));
-        row["cleanup_not_before"] = json!(timestamp_after(-60 * 60));
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200).set_body_json(json!([row])),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            response_json(response).await["error"]["code"],
-            "attachment_backup_conflict"
-        );
-        assert_eq!(server.received_requests().await.unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn accepts_finalize_retry_after_the_candidate_is_current() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("current", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = response_json(response).await;
-        assert_eq!(body["objectState"], "current");
-        assert_eq!(body["wasFinalized"], false);
-        assert_eq!(server.received_requests().await.unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn rejects_storage_integrity_mismatch_without_finalizing() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("reserved", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/info/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "size": 1234,
-                "content_type": "application/octet-stream",
-                "metadata": {
-                    "ciphertextSha256": "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
-                    "formatVersion": 1
-                }
-            })))
-            .mount(&server)
-            .await;
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        assert_eq!(server.received_requests().await.unwrap().len(), 2);
-    }
-
-    #[tokio::test]
-    async fn rejects_storage_content_mismatch_without_finalizing() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("reserved", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/info/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "size": 1234,
-                "content_type": "application/octet-stream",
-                "metadata": {
-                    "ciphertextSha256": CIPHERTEXT_SHA256,
-                    "formatVersion": 1
-                }
-            })))
-            .mount(&server)
-            .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/authenticated/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(vec![1_u8; 1_234]))
-            .mount(&server)
-            .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        assert_eq!(server.received_requests().await.unwrap().len(), 3);
-    }
-
-    #[tokio::test]
-    async fn throttles_parallel_storage_content_verification() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("reserved", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/info/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "size": 1234,
-                "content_type": "application/octet-stream",
-                "metadata": {
-                    "ciphertextSha256": CIPHERTEXT_SHA256,
-                    "formatVersion": 1
-                }
-            })))
-            .mount(&server)
-            .await;
-        let state = test_state(&server);
-        let _verification_slot = state
-            .attachment_verification_slots
-            .clone()
-            .try_acquire_owned()
-            .unwrap();
-
-        let response = test_router_with_state(state, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::SERVICE_UNAVAILABLE);
-        assert_eq!(server.received_requests().await.unwrap().len(), 2);
-    }
-
-    #[tokio::test]
-    async fn rejects_spoofed_storage_metadata_when_object_bytes_do_not_match() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("reserved", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/info/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "metadata": { "size": 1234, "mimetype": "application/octet-stream" },
-                "user_metadata": {
-                    "ciphertextSha256": CIPHERTEXT_SHA256,
-                    "formatVersion": 1
-                }
-            })))
-            .mount(&server)
-            .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/authenticated/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(vec![1; 1234]))
-            .mount(&server)
-            .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        assert_eq!(server.received_requests().await.unwrap().len(), 3);
-    }
-
-    #[tokio::test]
-    async fn reports_storage_verification_failures_as_unavailable() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("reserved", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/storage/v1/object/info/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(
-                ResponseTemplate::new(500)
-                    .set_body_json(json!({ "message": "storage-secret-must-not-leak" })),
-            )
-            .mount(&server)
-            .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/finalize",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
-        assert!(
-            !response_json(response)
-                .await
-                .to_string()
-                .contains("storage-secret")
-        );
-        assert_eq!(server.received_requests().await.unwrap().len(), 2);
-    }
-
-    #[tokio::test]
-    async fn reports_head_cas_conflict_without_leaking_upstream_details() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("ready", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        mount_rpc(
-            &server,
-            "promote_attachment_backup",
-            ResponseTemplate::new(409).set_body_json(json!({
-                "code": "40001",
-                "message": "current-key-secret"
-            })),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::PUT,
-                "/attachment-backups/head",
-                json!({
-                    "objectKey": object_key(),
-                    "expectedCurrentObjectKey": displaced_key()
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        let body = response_json(response).await;
-        assert_eq!(body["error"]["code"], "attachment_backup_conflict");
-        assert!(!body.to_string().contains("current-key-secret"));
-    }
-
-    #[tokio::test]
-    async fn returns_version_and_integrity_for_current_and_promoted_heads() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_current_attachment_backup",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "object_id": OBJECT_ID,
-                "version_ref": VERSION_REF,
-                "object_key": object_key(),
-                "ciphertext_sha256": CIPHERTEXT_SHA256,
-                "ciphertext_size_bytes": 1234,
-                "format_version": 1
-            }])),
-        )
-        .await;
-        let current = test_router(&server, true)
-            .oneshot(
-                Request::get(format!("/attachment-backups/head/{ATTACHMENT_REF}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(current.status(), StatusCode::OK);
-        let body = response_json(current).await;
-        assert_eq!(body["versionRef"], VERSION_REF);
-        assert_eq!(body["ciphertextSha256"], CIPHERTEXT_SHA256);
-
-        let promote_server = MockServer::start().await;
-        mount_rpc(
-            &promote_server,
-            "read_attachment_backup_by_key",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([object_row("ready", Some(CIPHERTEXT_SHA256))])),
-        )
-        .await;
-        mount_rpc(
-            &promote_server,
-            "promote_attachment_backup",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "current_object_id": OBJECT_ID,
-                "current_object_key": object_key(),
-                "current_version_ref": VERSION_REF,
-                "current_ciphertext_sha256": CIPHERTEXT_SHA256,
-                "displaced_object_id": null,
-                "displaced_object_key": null,
-                "was_promoted": true
-            }])),
-        )
-        .await;
-        let promoted = test_router(&promote_server, true)
-            .oneshot(json_request(
-                Method::PUT,
-                "/attachment-backups/head",
-                json!({ "objectKey": object_key(), "expectedCurrentObjectKey": null }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(promoted.status(), StatusCode::OK);
-        let body = response_json(promoted).await;
-        assert_eq!(body["currentVersionRef"], VERSION_REF);
-        assert_eq!(body["currentCiphertextSha256"], CIPHERTEXT_SHA256);
-    }
-
-    #[tokio::test]
-    async fn prepares_current_download_before_signing() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "prepare_attachment_backup_download",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "object_id": OBJECT_ID,
-                "object_key": object_key(),
-                "ciphertext_sha256": CIPHERTEXT_SHA256,
-                "ciphertext_size_bytes": 1234,
-                "format_version": 1,
-                "cleanup_not_before": timestamp_after(60 * 60)
-            }])),
-        )
-        .await;
-        Mock::given(method("POST"))
-            .and(path(format!(
-                "/storage/v1/object/sign/{ATTACHMENT_BACKUP_BUCKET}/{}",
-                object_key()
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_json(json!({
-                "signedURL": format!(
-                    "/object/sign/{ATTACHMENT_BACKUP_BUCKET}/{}?token=download-secret",
-                    object_key()
-                )
-            })))
-            .mount(&server)
-            .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/download",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = response_json(response).await;
-        assert_eq!(body["objectId"], OBJECT_ID);
-        assert_eq!(body["ciphertextSha256"], CIPHERTEXT_SHA256);
-        assert!(
-            body["signedUrl"]
-                .as_str()
-                .unwrap()
-                .contains("download-secret")
-        );
-        let requests = server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 2);
-        assert_eq!(
-            requests[0].url.path(),
-            "/rest/v1/rpc/prepare_attachment_backup_download"
-        );
-        assert!(requests[1].url.path().contains("/storage/v1/object/sign/"));
-    }
-
-    #[tokio::test]
-    async fn never_signs_a_download_for_an_unsafe_cleanup_window() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "prepare_attachment_backup_download",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "object_id": OBJECT_ID,
-                "object_key": object_key(),
-                "ciphertext_sha256": CIPHERTEXT_SHA256,
-                "ciphertext_size_bytes": 1234,
-                "format_version": 1,
-                "cleanup_not_before": timestamp_after(DOWNLOAD_URL_TTL_SECONDS + 60)
-            }])),
-        )
-        .await;
-
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/download",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::BAD_GATEWAY);
-        let requests = server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 1);
-        assert!(
-            !requests
-                .iter()
-                .any(|request| request.url.path().contains("/storage/v1/"))
-        );
-    }
-
-    #[tokio::test]
-    async fn never_signs_a_noncurrent_download() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "prepare_attachment_backup_download",
-            ResponseTemplate::new(409).set_body_json(json!({
-                "code": "55000",
-                "message": "not current"
-            })),
-        )
-        .await;
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/download",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        assert_eq!(server.received_requests().await.unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn schedules_exact_delete_identity_with_stable_replay_shape_without_storage() {
-        let delete_not_before = timestamp_after(24 * 60 * 60);
-        let created_server = MockServer::start().await;
-        mount_rpc(
-            &created_server,
-            "schedule_attachment_backup_deletion",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([scheduled_deletion_row(&delete_not_before, true)])),
-        )
-        .await;
-
-        let created_response = test_router(&created_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(created_response.status(), StatusCode::OK);
-        let created_body = response_json(created_response).await;
-        assert_eq!(
-            created_body,
-            json!({
-                "objectKey": object_key(),
-                "attachmentRef": ATTACHMENT_REF,
-                "versionRef": VERSION_REF,
-                "deleteRequestId": DELETE_REQUEST_ID,
-                "deleteFenceId": DELETE_FENCE_ID,
-                "deleteGeneration": 7,
-                "deleteNotBefore": delete_not_before
-            })
-        );
-        let created_requests = created_server.received_requests().await.unwrap();
-        assert_eq!(created_requests.len(), 1);
-        assert_eq!(
-            created_requests[0].url.path(),
-            "/rest/v1/rpc/schedule_attachment_backup_deletion"
-        );
-        assert_eq!(
-            serde_json::from_slice::<Value>(&created_requests[0].body).unwrap(),
-            json!({
-                "p_owner_user_id": OWNER,
-                "p_attachment_ref": ATTACHMENT_REF,
-                "p_version_ref": VERSION_REF,
-                "p_object_key": object_key(),
-                "p_delete_request_id": DELETE_REQUEST_ID
-            })
-        );
-
-        let replay_server = MockServer::start().await;
-        mount_rpc(
-            &replay_server,
-            "schedule_attachment_backup_deletion",
-            ResponseTemplate::new(200)
-                .set_body_json(json!([scheduled_deletion_row(&delete_not_before, false)])),
-        )
-        .await;
-        let replay_response = test_router(&replay_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(replay_response.status(), StatusCode::OK);
-        assert_eq!(response_json(replay_response).await, created_body);
-        let replay_requests = replay_server.received_requests().await.unwrap();
-        assert_eq!(replay_requests.len(), 1);
-        assert!(
-            replay_requests
-                .iter()
-                .all(|request| !request.url.path().contains("/storage/v1/"))
-        );
-
-        let dependency_server = MockServer::start().await;
-        mount_rpc(
-            &dependency_server,
-            "cancel_attachment_backup_deletion",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "outcome": "dependency_appeared",
-                "object_key": object_key(),
-                "was_cancelled": false
-            }])),
-        )
-        .await;
-        let dependency = test_router(&dependency_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete/cancel",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(dependency.status(), StatusCode::OK);
-        assert_eq!(
-            response_json(dependency).await,
-            json!({
-                "objectKey": object_key(),
-                "attachmentRef": ATTACHMENT_REF,
-                "versionRef": VERSION_REF,
-                "deleteRequestId": DELETE_REQUEST_ID
-            })
-        );
-        let dependency_requests = dependency_server.received_requests().await.unwrap();
-        assert_eq!(dependency_requests.len(), 1);
-        assert!(
-            dependency_requests
-                .iter()
-                .all(|request| !request.url.path().contains("/storage/v1/"))
-        );
-    }
-
-    #[tokio::test]
-    async fn distinguishes_delete_outcomes_from_generic_conflicts_without_storage() {
-        let dependency_server = MockServer::start().await;
-        mount_rpc(
-            &dependency_server,
-            "schedule_attachment_backup_deletion",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "outcome": "dependency_appeared",
-                "object_id": null,
-                "object_key": object_key(),
-                "delete_request_id": DELETE_REQUEST_ID,
-                "delete_fence_id": null,
-                "delete_generation": null,
-                "delete_not_before": null,
-                "was_created": false
-            }])),
-        )
-        .await;
-        let dependency = test_router(&dependency_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(dependency.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            response_json(dependency).await["error"]["code"],
-            "attachment_backup_dependency_appeared"
-        );
-        let dependency_requests = dependency_server.received_requests().await.unwrap();
-        assert_eq!(dependency_requests.len(), 1);
-        assert!(
-            dependency_requests
-                .iter()
-                .all(|request| !request.url.path().contains("/storage/v1/"))
-        );
-
-        let cancelled_server = MockServer::start().await;
-        mount_rpc(
-            &cancelled_server,
-            "schedule_attachment_backup_deletion",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "outcome": "cancelled",
-                "object_id": null,
-                "object_key": object_key(),
-                "delete_request_id": DELETE_REQUEST_ID,
-                "delete_fence_id": null,
-                "delete_generation": null,
-                "delete_not_before": null,
-                "was_created": false
-            }])),
-        )
-        .await;
-        let cancelled = test_router(&cancelled_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(cancelled.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            response_json(cancelled).await["error"]["code"],
-            "attachment_backup_delete_cancelled"
-        );
-        let cancelled_requests = cancelled_server.received_requests().await.unwrap();
-        assert_eq!(cancelled_requests.len(), 1);
-        assert!(
-            cancelled_requests
-                .iter()
-                .all(|request| !request.url.path().contains("/storage/v1/"))
-        );
-
-        let conflict_server = MockServer::start().await;
-        mount_rpc(
-            &conflict_server,
-            "schedule_attachment_backup_deletion",
-            ResponseTemplate::new(409).set_body_json(json!({
-                "code": "40001",
-                "message": "delete identity changed"
-            })),
-        )
-        .await;
-        let conflict = test_router(&conflict_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(conflict.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            response_json(conflict).await["error"]["code"],
-            "attachment_backup_conflict"
-        );
-        let conflict_requests = conflict_server.received_requests().await.unwrap();
-        assert_eq!(conflict_requests.len(), 1);
-        assert!(
-            conflict_requests
-                .iter()
-                .all(|request| !request.url.path().contains("/storage/v1/"))
-        );
-    }
-
-    #[tokio::test]
-    async fn cancels_exact_delete_identity_with_a_stable_idempotent_shape() {
-        let created_server = MockServer::start().await;
-        mount_rpc(
-            &created_server,
-            "cancel_attachment_backup_deletion",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "outcome": "cancelled",
-                "object_key": object_key(),
-                "was_cancelled": true
-            }])),
-        )
-        .await;
-        let created = test_router(&created_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete/cancel",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(created.status(), StatusCode::OK);
-        let created_body = response_json(created).await;
-        assert_eq!(
-            created_body,
-            json!({
-                "objectKey": object_key(),
-                "attachmentRef": ATTACHMENT_REF,
-                "versionRef": VERSION_REF,
-                "deleteRequestId": DELETE_REQUEST_ID
-            })
-        );
-        let requests = created_server.received_requests().await.unwrap();
-        assert_eq!(requests.len(), 1);
-        assert_eq!(
-            requests[0].url.path(),
-            "/rest/v1/rpc/cancel_attachment_backup_deletion"
-        );
-        assert_eq!(
-            serde_json::from_slice::<Value>(&requests[0].body).unwrap(),
-            json!({
-                "p_owner_user_id": OWNER,
-                "p_attachment_ref": ATTACHMENT_REF,
-                "p_version_ref": VERSION_REF,
-                "p_object_key": object_key(),
-                "p_delete_request_id": DELETE_REQUEST_ID
-            })
-        );
-        assert!(
-            requests
-                .iter()
-                .all(|request| !request.url.path().contains("/storage/v1/"))
-        );
-
-        let replay_server = MockServer::start().await;
-        mount_rpc(
-            &replay_server,
-            "cancel_attachment_backup_deletion",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "outcome": "cancelled",
-                "object_key": object_key(),
-                "was_cancelled": false
-            }])),
-        )
-        .await;
-        let replay = test_router(&replay_server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete/cancel",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(replay.status(), StatusCode::OK);
-        assert_eq!(response_json(replay).await, created_body);
-        let replay_requests = replay_server.received_requests().await.unwrap();
-        assert_eq!(replay_requests.len(), 1);
-        assert!(
-            replay_requests
-                .iter()
-                .all(|request| !request.url.path().contains("/storage/v1/"))
-        );
-    }
-
-    #[tokio::test]
-    async fn reports_delete_cancellation_after_gc_as_too_late() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "cancel_attachment_backup_deletion",
-            ResponseTemplate::new(409).set_body_json(json!({
-                "code": "55006",
-                "message": "gc already owns the object"
-            })),
-        )
-        .await;
-        let response = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete/cancel",
-                delete_request_body(),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            response_json(response).await["error"]["code"],
-            "attachment_backup_delete_too_late"
-        );
-    }
-
-    #[tokio::test]
-    async fn requires_pro_and_rejects_bad_inputs_before_upstream_calls() {
-        let server = MockServer::start().await;
-        let non_pro = test_router(&server, false)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/download",
-                json!({ "objectKey": object_key() }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(non_pro.status(), StatusCode::FORBIDDEN);
-        assert_eq!(
-            response_json(non_pro).await["error"]["code"],
-            "subscription_required"
-        );
-
-        for path in [
-            "/attachment-backups/delete",
-            "/attachment-backups/delete/cancel",
-        ] {
-            let non_pro_delete = test_router(&server, false)
-                .oneshot(json_request(Method::POST, path, delete_request_body()))
-                .await
-                .unwrap();
-            assert_eq!(non_pro_delete.status(), StatusCode::FORBIDDEN);
-            assert_eq!(
-                response_json(non_pro_delete).await["error"]["code"],
-                "subscription_required"
-            );
-        }
-
-        let invalid_key = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/download",
-                json!({ "objectKey": "../other-user/object" }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(invalid_key.status(), StatusCode::BAD_REQUEST);
-
-        let invalid_hash = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/upload-grant",
-                json!({ "objectKey": object_key(), "ciphertextSha256": "ABC" }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(invalid_hash.status(), StatusCode::BAD_REQUEST);
-
-        let unknown_field = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete",
-                json!({
-                    "objectKey": object_key(),
-                    "attachmentRef": ATTACHMENT_REF,
-                    "versionRef": VERSION_REF,
-                    "deleteRequestId": DELETE_REQUEST_ID,
-                    "ownerUserId": "attacker"
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(unknown_field.status(), StatusCode::UNPROCESSABLE_ENTITY);
-
-        let noncanonical_request_id = test_router(&server, true)
-            .oneshot(json_request(
-                Method::POST,
-                "/attachment-backups/delete/cancel",
-                json!({
-                    "objectKey": object_key(),
-                    "attachmentRef": ATTACHMENT_REF,
-                    "versionRef": VERSION_REF,
-                    "deleteRequestId": "{44444444-4444-4444-8444-444444444444}"
-                }),
-            ))
-            .await
-            .unwrap();
-        assert_eq!(noncanonical_request_id.status(), StatusCode::BAD_REQUEST);
-        assert!(server.received_requests().await.unwrap().is_empty());
-    }
-
-    #[tokio::test]
-    async fn rejects_oversized_or_origin_injecting_ledger_responses() {
-        let server = MockServer::start().await;
-        mount_rpc(
-            &server,
-            "read_current_attachment_backup",
-            ResponseTemplate::new(200)
-                .set_body_string("x".repeat(MAX_BACKUP_RPC_RESPONSE_BYTES + 1)),
-        )
-        .await;
-        let oversized = test_router(&server, true)
-            .oneshot(
-                Request::get(format!("/attachment-backups/head/{ATTACHMENT_REF}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(oversized.status(), StatusCode::BAD_GATEWAY);
-
-        let second_server = MockServer::start().await;
-        mount_rpc(
-            &second_server,
-            "read_current_attachment_backup",
-            ResponseTemplate::new(200).set_body_json(json!([{
-                "object_id": OBJECT_ID,
-                "version_ref": VERSION_REF,
-                "object_key": "https://attacker.example/secret",
-                "ciphertext_sha256": CIPHERTEXT_SHA256,
-                "ciphertext_size_bytes": 1234,
-                "format_version": 1
-            }])),
-        )
-        .await;
-        let malformed = test_router(&second_server, true)
-            .oneshot(
-                Request::get(format!("/attachment-backups/head/{ATTACHMENT_REF}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(malformed.status(), StatusCode::BAD_GATEWAY);
-        assert!(
-            !response_json(malformed)
-                .await
-                .to_string()
-                .contains("attacker.example")
-        );
-    }
-}
+mod tests;
